Move dashboards components into dashboard directory.
