+++ graylog2-web-interface/src/views/components/DashboardActionsMenu.tsx
@@ -41,7 +41,7 @@ import {
 } from 'views/logic/views/pluggableSaveViewFormHandler';
 import useSaveViewFormControls from 'views/hooks/useSaveViewFormControls';
 
-import DashboardPropertiesModal from './views/DashboardPropertiesModal';
+import DashboardPropertiesModal from './dashboard/DashboardPropertiesModal';
 import BigDisplayModeConfiguration from './dashboard/BigDisplayModeConfiguration';
 
 const _isAllowedToEdit = (view: View, currentUser: User | undefined | null) => isPermitted(currentUser?.permissions, [ViewPermissions.View.Edit(view.id)])
