+++ graylog2-web-interface/src/views/components/dashboard/DashboardPropertiesModal.test.tsx
