+++ graylog2-web-interface/src/views/components/dashboard/DashboardsOverview/DashboardActions.tsx
