+++ graylog2-web-interface/src/views/components/dashboard/DashboardsOverview/index.ts
