+++ graylog2-web-interface/src/views/components/searchbar/saved-search/SearchActionsMenu.tsx
@@ -34,7 +34,7 @@ import EntityShareModal from 'components/permissions/EntityShareModal';
 import useCurrentUser from 'hooks/useCurrentUser';
 import * as ViewsPermissions from 'views/Permissions';
 import type User from 'logic/users/User';
-import ViewPropertiesModal from 'views/components/views/DashboardPropertiesModal';
+import ViewPropertiesModal from 'views/components/dashboard/DashboardPropertiesModal';
 import { loadAsDashboard, loadNewSearch } from 'views/logic/views/Actions';
 import IfPermitted from 'components/common/IfPermitted';
 import {
